refactor!: move `NodeEventHandler` into `lib.rs`
